IFIX-374-changes related to the expense

+++ frontend/mgramseva/lib/model/connection/property.dart
@@ -37,7 +37,7 @@ class Property {
   @JsonKey(name: "units")
   List<Units>? units;
   @JsonKey(name: "landArea")
-  int? landArea;
+  double? landArea;
   @JsonKey(name: "propertyType")
   String? propertyType;
   @JsonKey(name: "noOfFloors")
